### 2025-10-07 #9165 @editorialbot
[REVIEW]: My awesome paper
**Submitting author:** @diehlpk (<a href="http://orcid.org/0000-0003-3922-8419">Patrick Diehl</a>)
**Repository:** https://github.com/diehlpk/my_awesome_joss_paper/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
