### 2025-10-07 #9165 @willgearty
👋🏼 @diehlpk @danielskatz this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
