### 2025-10-07 #9165 @danielskatz
Ok - I think this is now ready to be accepted - I'm happy with it

thanks for all your hard work @diehlpk!
